requirements.txt
python-novaclient>=2.22.0
